Addition of install.bat and .vsconfig file to automate the setup process of our project (#11)
* This file has components for Visual studio

This file has component ID's for Visual studio build tools to build wheels for ChromaDB

* Batch file for hassle free replication of project 

This batch script automates the setup process for this project, ensuring that Python 3.12.0 is installed and configured correctly. It also installs the Visual Studio Build Tools for building wheels and installs specific dependencies using pip. Just run the script and let it handle the setup for you!